+++ Assets/Graphics/Prototype/CheckerGray.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 785391246c69374308d3bb1b244a9464
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
