+++ Assets/Graphics/Prototype/CheckerWhite.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cb9e414379ffd89ab8f2ba6a99ab8bc0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
