+++ Assets/Physics/NoFriction.physicMaterial
@@ -1,11 +1,14 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!62 &6200000
-PhysicsMaterial2D:
+--- !u!134 &13400000
+PhysicMaterial:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_Name: NoFriction
-  friction: 0
+  dynamicFriction: 0
+  staticFriction: 0
   bounciness: 0
+  frictionCombine: 0
+  bounceCombine: 0
