+++ Assets/Physics/NoFriction.physicMaterial.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 96c03acd24e805e4b89fbe84d1477b31
+guid: f5c0b7fc47599ac8ba25ae3d8115162d
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 6200000
+  mainObjectFileID: 13400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
